Make Additional Function Capture Precise
Summary:
Release Notes: none.

This fixes #1093. At a high level, captures are currently needed by the additional function only if they are modified.

Implementation-wise we tentatively create all bindings with callbacks to visit them and delete them to make sure that after Additional Functions are visited we have a precise set of captures.

Functionality of the PR is fully there but I would like to refactor it in some way because I feel like it can be cleaner.
Closes https://github.com/facebook/prepack/pull/1107

Differential Revision: D6290712

Pulled By: cblappert

fbshipit-source-id: 8b812e28f83c2cb216c01eb52a10f6cb0873ea8d