Do not dead-code eliminate do-while in residual fns
Summary:
Followup to #1061; I thought of another case where it was too aggressive.

Do-while loops with always-false conditions can't be replaced by their body because their body might contain a break or continue, which is illegal (or will have different semantics) if the loop becomes a normal block. In principle the body of the loop could be walked to find those, but frankly I doubt it will come up enough to be worth it.
Closes https://github.com/facebook/prepack/pull/1145

Differential Revision: D6303312

Pulled By: hermanventer

fbshipit-source-id: 0d677076fea6f225573a36fbbc9df1a0bebd959a