Remove modules with static functions from the big Flow cycle, part 1
Summary:
In our current setup a big group of beefy modules form an import cycle and Flow needs to analyze the entire cycle together, which makes type checking with Flow very slow for us.

The length of the cycle has also crept up from 52 to 62 in recent months, despite an explicit check-in gate designed to prevent this.

Reducing the cycle is not trivial, since the degree of connectivity from one node to another is quite high and because there is no tooling for finding the least connected node and removing that from the cycle.

So, with this request, I'm starting down the road of introducing an abstract class for each module in the methods directory and then changing the module to into a class where every exported static function becomes an instance method of an implementation class with a singleton instance. The instance is found via the singletons.js module, which does not have dependencies on the implementation modules. The properties of the singleton module are initialized, early on, in the default function of the contruct_realm.js module.

As of yet, the cycle count has not decreased, but two beefy nodes have been exchanged for two smaller nodes.
Closes https://github.com/facebook/prepack/pull/1139

Differential Revision: D6314194

Pulled By: hermanventer

fbshipit-source-id: f9656b84d6d056006f0734da295523f8449d9ef2