Join properties with get/set
Summary:
Release notes: Provides support for joining properties that have get and set methods. Also now fully support properties that are only defined in one branch of an unknown conditional.

The main thing that happens here is that descriptors that result from joins now track the join condition and the original descriptors, so that the serializer can emit conditional code for defining the entire property, which allows it to also conditionally omit properties altogether.

Issue #1015
Closes https://github.com/facebook/prepack/pull/1116

Differential Revision: D6331039

Pulled By: hermanventer

fbshipit-source-id: 42d635d37b9880398cc4e039f2d59a9a5d03cec0