Send multiple breakpoints in one debugger request
Summary:
Release note: none

Previously, the DebugAdapter sends breakpoint requests one by one to Prepack. This works for the CLI because breakpoints are set one at a time. For an IDE, multiple breakpoints are sent in a single request upon startup and a single response with all the breakpoints set is expected. This PR sends allows sending multiple breakpoints in a single request to Prepack and a single response is received.
Closes https://github.com/facebook/prepack/pull/1160

Differential Revision: D6351382

Pulled By: JWZ2018

fbshipit-source-id: 5b84c22d257da4b943dc8ef4856fefd006e3921e