Avoid unused property reads on simple objects
Summary:
Release note: Fixes #1179

When a property is written to object, the rather convoluted logic for checking if it is allowed and if a setter is involved end up asking the object three times if it already has such a property.

In the case of a simple partial object, this caused three useless references to the property to appear in the generated code, since the absence of a property is regarded as OK and produces a temporal abstract value that just reads from the object at runtime.

Fixed this by adding a special case when writing to a missing property on a simple partial object.
Closes https://github.com/facebook/prepack/pull/1183

Differential Revision: D6399970

Pulled By: hermanventer

fbshipit-source-id: b964ee399583f3f4f30dc2f1930e06a856952635