Treat first operand of a logical operation as a condition
Summary:
Release note: Partially fixes #1182

When dealing with x || y and x && y where x is an abstract value, we can avoid evaluating x if it is known to be false in the first case and true in the second. When x is being refined with the path condition, it helps to know that the refined result will only be checked for truthiness. Duly add a call to Environment.GetConditionValue to make this happen.
Closes https://github.com/facebook/prepack/pull/1184

Differential Revision: D6420018

Pulled By: hermanventer

fbshipit-source-id: ec45808b0235cda896634c09940a8d20b318eb48