Fix lazy object identifier inside ConditionalAssignmentBranch
Summary:
Release Note: Fix lazy object identifier inside ConditionalAssignmentBranch

Fix #1187. The lazy object identifier logic explicitly checks if the serialization body is lazy object initializer; however, it's possible that object is serializing inside a nested ConditionalAssignmentBranch which the logic will fail.
To fix this issue, I added a "parentBody" field to all serializerBody which chains the serialized body as a chain so that we can check the nested children body cases.
Also, add the remaining 3 hook methods in proxy hook which failed because of this bug.
Closes https://github.com/facebook/prepack/pull/1198

Differential Revision: D6427510

Pulled By: yinghuitan

fbshipit-source-id: 39bde34bd508045b9c4dd109538a905a203a0dfe