Do not let thrown objects escape from delayed modules
Summary:
The logic for delaying modules that fail to prepack did the wrong thing with modules that always throw, by leaking the exception object while backing out the effects of the require call.

To keep things simple all modules that end abruptly are now delayed (when the option is specified) and the exception is discarded.

Eventually, we should just get rid of the delay logic, but for now it seems like useful backup. Feel free, however, to insist that we get rid of it right away.
Closes https://github.com/facebook/prepack/pull/1203

Differential Revision: D6442540

Pulled By: hermanventer

fbshipit-source-id: e3a609af8c37cbe4ac9a8e96fb5c1950c886b435