Update test262 failure threshold for ES5 and ES6 ids
The ES5 and ES6 ids are legacy flags from Test262 and they are mostly used to flag features implemented by ES5.1 or ES6 without any update in consecutive spec editions. They contain a matching reference for the section number of the respective spec edition. New tests are supposed to have only the esid tag for matching sections by their respective HTML anchor ids.

With the recent updates on Test262 this number is changing so the threshold for failures on tests with the es5id and es6id tags needs to reflect this same change.