Update expected number of test262 passes
Summary:
It seems that the numbers were meant for another version of the test262 repository.

The original pull request updated the test262 repository label, but the bot that synchronizes GitHub with our internal repo removed that update. That also causes the new test runner to fail, so disable that until we update the label properly.
Closes https://github.com/facebook/prepack/pull/1215

Differential Revision: D6477069

Pulled By: hermanventer

fbshipit-source-id: ed17b12c1e81d51465f9c4b76adf684961390fb0