Never mutate the state of a leaked object
When an object leaks, we need to store its initial state somehow so that
we can pass it in that state to the function.

In previous attempts I serialized it as an empty object and then emitted
statements to the generator that recreates it. This doesn't work when
only one branch leaks and the other one doesn't since we don't have a way
to serialize the complete object that didn't leak.

It is also tricky to replicate a complete object serialization by just
adhoc emitting statements.

The realization I had was that once an object leaks, you can't really make
many assumptions about its future state. There are many things that can
invalidate what you think you know about the object. E.g. getters/setters
could have been added. Any time you call another abstract function, that
might have mutated any reachable object.

While we could reason about mutations to non-configurable, writable values
over time, this is a pretty rare configuration.

It simplifies the model if we just treat a leaked object as completely
unknown except for non-configurable descriptors and possibly non-writable
values.

That way serializing the initial state of leaked objects is easy.
We just get that for free.