Implement to.js privately
Summary:
Move the implementations of the to methods behind an interface so that we can reduce the Flow cycle length (as well as avoiding increasing it when a planned future request adds a new import to to.js).

This is a mechanical refactor with no new functionality.
Closes https://github.com/facebook/prepack/pull/1219

Differential Revision: D6494524

Pulled By: hermanventer

fbshipit-source-id: b5dbe45155e0276a9d42f2b26a90bb9b220ac1ea