Never mutate value of leaked binding
Do the same thing as we now do with objects. Instead of storing the
initial value, we simply never mutate a binding once it has leaked.

This avoids problems when it can have been mutated in unexpected places
after the original leak. It also avoids complexities dealing with
branches where it leaks in one but not the other.