Leak any values assigned to a leaked object or binding
When an object leaks or a binding leaks. That means that some external
code can essentially read that at any time or define a setter on leaked
objects.

We could be clever about tracking when this could happen.

The simple solution is to simply also leak any value that gets assigned to a
leaked object or binding.