Only leak bindings that are actually referenced by a leaked function
To do this, we need to traverse the function body to find them.

Ideally we'd have a more general cache for function info that can be
used by both the serializer and this but I'll keep it separate for now.

We only need to leak the value if it can be read. We only need to leak
the binding if it is written to.