Error if an abstract function that might throw is called from within a try
This is a follow up to #1142. If we call an unknown function it might throw
which will take a different control Flow. #1264 is the follow up to do this
properly but since we don't have a use case right now, we can just issue
a recoverable error if this happens.