Provide for intrinsic abstract arrays
Summary:
Release note: none

Add a do-while test where the termination condition involves the length of an abstract array.

Since we did not have abstract arrays, this is mainly about adding those. Also, there was no support for getting a property from an abstract object that is not template based, so added that too. Also added a special case to return an abstract value of type NumberValue if the abstract object is known to be an array and the property name is known to be "length".

Finally, fixed a typing error in createAbstractFunction. It turns out, ahem, that Flow was not confused at all, just obtused.
Closes https://github.com/facebook/prepack/pull/1297

Differential Revision: D6633451

Pulled By: hermanventer

fbshipit-source-id: 23302999002ca35d3141a1666afd76c3b9208de7