Fix website dropdowns
Summary:
Issue https://github.com/facebook/prepack/issues/1296.

The previous version of Tether Select that was used in the website had `z-index: 100;` (still can check it in `gh-pages` branch) but in the version that we use from CDN, in order to reduce the size of Prepack repo, there is no `z-index`. So I've added it to the website styles.
Closes https://github.com/facebook/prepack/pull/1298

Differential Revision: D6638548

Pulled By: NTillmann

fbshipit-source-id: 1fb05a37d7bd83a71936ac4aa4ed9a138825f032