Unescape uniqueSuffix for test-runner output
Summary:
Release Note: none.

Currently, if a test failed, the output of the prepacked code is augmented with uniqueSuffix "_unique_27787". This makes the code very hard to read and not efficient for troubleshooting.
Unescape uniqueSuffix for the Prepacked output to fix this issue.

Furthermore, is there any deep technical reason that we need to use uniqueSuffix? I know it is supposed for test-runner fixed point calculation, but I do not see any technical reason it is needed. For example, I removed uniqueSuffix, all tests still pass. Let me know if you want me to remove it completely from test runner.
Closes https://github.com/facebook/prepack/pull/1294

Differential Revision: D6638856

Pulled By: yinghuitan

fbshipit-source-id: 3887402fe391f67981948cfdd6ab3ad878c307e5