Detect abstract recursion
Summary:
This is the first step towards handling recursive functions where the recursive calls as guarded by abstract conditions. Prior to this, Prepack failed with a stack overflow. Now it detects such calls and computes a fixed point to model its state.

As yet, the resulting state cannot be serialized, so currently such recursive functions fail with a generic diagnostic message.

The basic idea is to record the calling arguments and location of each active call that occurs in a guarded block. If the same calling location is reached again, but with an additional guard, this is likely an recursion what will not terminate, so the arguments to the new call are widened and the recording is updated. At some point the recorded arguments contain the widened arguments and a fixed point is reached. At that point the nested call is declared as abstract recursive and it is replaced with an abstract value with domains all set to Top. Eventually, a set of effects is computed for the outer most call.

In future pull requests, these effects will be used to generate a residual function that can be called at runtime (so that recursion can be used).
Closes https://github.com/facebook/prepack/pull/1311

Differential Revision: D6664532

Pulled By: hermanventer

fbshipit-source-id: 0f8377f564a0fed6372e0854a90f41c8fd9d5ee8