Error if an abstract function that might throw is called within a try
Summary:
This is a follow up to #1142. If we call an unknown function it might throw which will take a different control flow. #1264 is the follow up to do this properly but since we don't have a use case right now, we can just issue a recoverable error if this happens.
Closes https://github.com/facebook/prepack/pull/1284

Differential Revision: D6669732

Pulled By: sebmarkbage

fbshipit-source-id: 32ff40d4afba589ce60193b01034dfb8bc764378