Don't error calling possibly non-functions in pure functions
As outlined in https://github.com/facebook/prepack/issues/1264 we know
that it is safe to throw in an pure function since then the return value
will be ignored and therefore its state doesn't matter. The only time
it matters, is if there is a try/catch around it which we already warn
about.

This means that we don't have to error when calling a non-function.

I also plan to add more of these cases like this. So I took the liberty to
also move this category of tests into a separate folder.