Let $Get succeed on abstract values
Property access is not safe on abstract object values because they could be
getters with side-effects or might throw.

However, this is safe in a pure function context. We just have to make sure
we leak the object that might get passed to a getter.

It is also safe to coerce a non object abstract value to an object, since
all it can do is throw.