Add evaluateWithPossibleThrowCompletion
I'm going to need to add a lot of these things that might throw so I wanted
to start preparing the infrastructure for fixing #1264.

The idea is that code inside this scope might emit temporal operations that
might throw. In the future we can wrap these in a try/catch to extract a
caught error and a flag. These can be used to generate a
PossiblyNormalCompletion.

However, for now I just keep the existing strategy of erroring in this case
while I add more callers of this helper.