Support Abstract Property Access in Pure Functions
Summary:
Builds on top of #1321

Lets $Get succeed on abstract values. Property access is not safe on abstract object values because they could be getters with side-effects or might throw.

However, this is safe in a pure function context. We just have to make sure we leak the object that might get passed to a getter.

It is also safe to coerce a non-object abstract value to an object, since all it can do is throw.
Closes https://github.com/facebook/prepack/pull/1322

Differential Revision: D6734870

Pulled By: sebmarkbage

fbshipit-source-id: bfe7a61dc312577ec4cc8926d3191dcee4e28b61