Add evaluateWithPossibleThrowCompletion
Summary:
I'm going to need to add a lot of these things that might throw so I wanted to start preparing the infrastructure for #1264. See that issue for more info on the approach.

The idea is that code inside this scope might emit temporal operations that might throw. In the future we can wrap these in a try/catch to extract a caught error and a flag. These can be used to generate a PossiblyNormalCompletion.

However, for now I just keep the existing strategy of erroring in this case while I add more callers of this helper.
Closes https://github.com/facebook/prepack/pull/1326

Differential Revision: D6735259

Pulled By: sebmarkbage

fbshipit-source-id: c16a4ed2d4eaf9120dc38d60a4922c7cab0ba59c