Deal with composition of saved effects
Summary:
Release note: partially fixes issue #1266

When function that conditionally throws an exception is called on the normal of path of an earlier conditional exception, the exceptional state stored in Realm.savedCompletion was not correctly composed when the function returned. In particular, the savedEffects from the function was just thrown away. This pull request preserves that state and deals with the subtle issue that arise from undoing and redoing saved effects. In particular, effects should never be shared and undo/redo should be compositional.

The remaining issue with #1266 is unrelated and will have to wait.
Closes https://github.com/facebook/prepack/pull/1332

Differential Revision: D6750548

Pulled By: hermanventer

fbshipit-source-id: 72bdabddc15765101e7451faa3655a4867ee6289