Do not visit ES2015 class prototypes that are null values
Summary:
Release notes: none

This fixes the serialization error in `/test/error-handler/class.js`. We don't want to serialize the value of `null` out for the prototype when we serialize ES2015 class syntax that extends a value that might have a `null` prototype anyway. This change fixes the test as expected and no errors are printed.
Closes https://github.com/facebook/prepack/pull/1335

Reviewed By: hermanventer

Differential Revision: D6751003

Pulled By: trueadm

fbshipit-source-id: e9586f4ea97464f1254d48e1aae0048f385897fd