Add a --check option
Summary:
Release note: Added a new option, --check, which instructs prepack to run over all reachable code and to report all of the applicable diagnostics.

So far the option is hooked up and it ensures that all module factories are run, but residual functions are not yet checked. That is a bit more complicated and is best left to separate pull requests.
Closes https://github.com/facebook/prepack/pull/1342

Differential Revision: D6774867

Pulled By: hermanventer

fbshipit-source-id: 857fa8ce8f7d1372930462a85e8f62480328752d