Preserve union types in model object properties
Summary:
Release note: Fixed problems with __abstractOrUndefined and friends

The code for making model property values intrinsic, did not have special knowledge of abstract unions and hence destroyed them. Likewise the code for obtaining an temporal reference to such a property did not preserve the union in the resulting derived value.

Added special case handling for this and added more special cases in the type inspection properties of AbstractValue.
Closes https://github.com/facebook/prepack/pull/1355

Differential Revision: D6816045

Pulled By: hermanventer

fbshipit-source-id: 580f269e34f97a1d1a278d7386c7c2ae2a7f8740