Hook up partial evaluator as a post processing step for the serializer.
Summary:
Release note: none

Provide an option to run the partial evaluator over the output from the serializer when running serializer tests.

Fixed up a few bugs that came to light from the new tests.

Exclude a few test cases that are awkward to run correctly in this configuration.
Closes https://github.com/facebook/prepack/pull/1349

Differential Revision: D6821423

Pulled By: hermanventer

fbshipit-source-id: aaae155a8be204647946a8abe64c8b03d5e3129e