Add more React Relay mocks
Summary:
I need them for the UFI.

(Side note: the "not a function" message I got otherwise was pretty terrible. Would never have figured out it was due to a missing definition without a debugger. We should improve it to show the actual name of the function that doesn't exist.)
Closes https://github.com/facebook/prepack/pull/1366

Reviewed By: trueadm

Differential Revision: D6837200

Pulled By: gaearon

fbshipit-source-id: b9101d9f525dd37bce24eeefbd81f945f72b3917