Handle abstract `toString` casting in pure additionalFunctions
Summary:
Release notes: none

When attempting to do `String(props.title)` on an abstract value, it currently throws a FatalError: `PP0001: This operation is not yet supported on abstract value`. Ideally, when we're in pure evaluation we can mark this as dirty too like we do other abstract values. I've added a test case to show this failing.
Closes https://github.com/facebook/prepack/pull/1361

Reviewed By: sebmarkbage

Differential Revision: D6849468

Pulled By: trueadm

fbshipit-source-id: c75d2f8c869e1f892ae0a330c3c2bd55ca2a9e6e