Resurrecting "require optimization" with dependency indirections.
Summary:
Release notes: Calls to the require-function within modules are (again) replaced by direct referenced to the prepacked exported object, if the required module has been prepacked.

This addresses #1101, and is achieved by additional pattern matching for the new __d and require call patterns emitted by the Metro bundler.

Verified with an internal sizable benchmark that this affects as expected a large number of require calls.
Closes https://github.com/facebook/prepack/pull/1369

Differential Revision: D6847036

Pulled By: NTillmann

fbshipit-source-id: b02b2c1db2b86899eb4bc72e386e8a6fafeb3da6