Analyze residual functions
Summary:
Release note: Provide a way to analyze residual functions for Prepack errors

Since the analysis takes a while, the --check option now takes start and count parameters so that a subrange of functions can be analyzed.

Each function is analyzed only in the context of the heap as it was at the end of global code execution and the arguments supplied to the analysis are all simple partial abstract objects, so that the analysis of a function does not come to screeching halt every time a parameter is used as the object of a member expression or as the argument of an operator.

A number of tweaks to the handling of simple objects will be in a follow up pull request.
Closes https://github.com/facebook/prepack/pull/1379

Differential Revision: D6876045

Pulled By: hermanventer

fbshipit-source-id: 4d192cc6cf222b3986114362c5642a949f292845