Do not ever ignore the set constructor argument
Summary:
Release note: none

It does not seem likely that code that passes an argument to the Set constructor when targeting jsc-600-1-4-17 actually intended for it to be ignored, so just use it when running Prepack, even if this would deviate from the behavior of the unprepacked code.

This also helps to deal with code that was supposed to use a polyfill, but that runs without the polyfill at Prepack time, because the module that loads the polyfill has been delayed.
Closes https://github.com/facebook/prepack/pull/1405

Differential Revision: D6915835

Pulled By: hermanventer

fbshipit-source-id: 123e25476f9640be7718f2dcba03ad5b5ea3577c