Refine ReactRelay mocks and add a `reactHints` tracking WeakMap
Summary:
Release notes: none

This PR adds `reactHints` property to AbstractValue and refines the ReactRelay mocks. `reactHints` allow us to do more with the React reconciler, without having to bail-out on encountering ReactRelay containers. So we now create the containers with a NativeFunction that allows us to finely create the `AbstractObjectValue`s back with the hints attached.

Without this change, the reconciler is unable to understand how to deal with ReactRelay container wrapped components (as they're left as abstract). We want to the containers abstract though, but we want to able to continue to process all the components in the tree. The hints provides special casing for making this work.

Furthermore, I've added tests that now properly incorporate ReactRelay and a relevant stub for tests that don't need the runtime in tests.
Closes https://github.com/facebook/prepack/pull/1418

Differential Revision: D6948627

Pulled By: trueadm

fbshipit-source-id: a6fb3690d8d9b472322ae1a8752f5a71a0747b52