Properly supporting Object.[freeze|sealed|preventExtensions] in serializer.
Release notes: Support for Object.freeze, Object.seal, and Object.preventExtensions.

The interpreter always did the right things, but the serializer ignored the $Extensible internal slot.
This is now fixed, and the serializer generates minimal code that first creates an object with simple assignments or an object literal, followed by a call to Object.freeze and Object.seal if needed.
This addresses #953 and #585.