Non destructive update of unions that get derived
Summary:
Release note: fixes issue #1411

When an abstract value that is of kind "abstractConcreteUnion" is specified as the value of an environmental object property, a derived value is created to capture the temporal value of the environment. This value must be turned back into a union, in order to preserve the original behavior.

This was done by mutating the original union, which is all sort of wrong and wreak havoc when the environmental property is accessed more than once.
Closes https://github.com/facebook/prepack/pull/1425

Differential Revision: D6954958

Pulled By: hermanventer

fbshipit-source-id: 615521b380afbb30120355c10d017e08f4c35af3