Properly supporting Object.[freeze|sealed|preventExtensions] in serializer.
Summary:
Release notes: Support for Object.freeze, Object.seal, and Object.preventExtensions.

The interpreter always did the right things, but the serializer ignored the $Extensible internal slot.
This is now fixed, and the serializer generates minimal code that first creates an object with simple assignments or an object literal when possible, followed by a call to Object.freeze / Object.seal/ Object.preventExtensions if needed.
This addresses #953 and #585.
Closes https://github.com/facebook/prepack/pull/1436

Differential Revision: D6958527

Pulled By: NTillmann

fbshipit-source-id: 3ab03595d5c033e6cddd6eb9a5947d7ef9f3d6e0