Fixing serialization order in presence of temporal dependencies
Release notes: Fixing ordering issues in serializer.

This is a refactoring of how serialization is done in the presence of temporal dependencies.

Before, there was a global set of declared identifiers that were already processed.
However, this was an imprecise approach, as the serializer can process values out-of-order for a number of reasons (when a value is used from multiple scopes and its definition needs to go to a particular body, and for lazy objects, and for lazy initializations (that get moved into residual functions)).
After this change, the serializer keeps track of declared-value sets for particular (nested) generators, and only merges them back into their parent generators when the generator is "done". This fixes the behavior of the waiting-logic to defer references to a declared value until after the generator that declares the value is "done".

Furthermore, when a value is used from multiple scopes, we used to emit that value into the body that corresponds to the common ancestor of all scopes; however, if that value was dependent on another value that was only declared in a nested generator that is stilled being processed (not yet done), then the code for that value would effectively end up *before* the nested generator with the dependent value declaration gets embedded in the code.
To address this issue, the algorithm in `getTarget` now checks if any of the scopes of a value intersects with the set of bodies that are not done yet contain a relevant dependency declaration. And if so, pick the most nested such body.

Adding test cases.