Generate Object.assign() calls for all arguments after first partial
We used to only generate residual Object.assign() calls for partial arguments.

The problem is, this breaks the case like Object.assign({}, partial, nonPartial).
We would incorrectly skip over nonPartial in this case.

With this fix, we generate Object.assign() for each argument after (and including) the first partial,
regardless of whether those arguments themselves are partial or not.