Make it so that !!(!x ? null : {}) => x
Summary:
Release note: Add another special case to the simplifier

Since !x on the path causes !!(!x ? null : {}) to simplify to false, it is not OK to explore a path with condition !x while exploring a path with condition !!(!x ? null : {}) because the existing path condition will be refined with the new condition, which will cause the simplifier to push a false condition on the path, which we disallow since such a path is clearly infeasible.

The null and {} in the above are illustrative only. Any expressions that coerce to false and true, respectively will do.
Closes https://github.com/facebook/prepack/pull/1438

Differential Revision: D7008070

Pulled By: hermanventer

fbshipit-source-id: 211a80751482bc9849c66b01ed223fd7646d10d3