Fatal or leak on mutations to Object.assign target
Object.assign() target must be in its final form during the call. Otherwise we can get bad output.

To ensure this, we add makeFinal() and isFinalObject() methods to the ObjectValue.
Once an object is "final", it cannot be changed.

Practically, this means that in the impure mode, we allow Object.assign() with partial simple sources but we forbid further mutations on the target.
They are treated as fatal errors (because without a way to "snapshot" states, it wouldn't be safe).

In pure mode, we leak the target on the first mutation.