Make sure to use derived reference in full invariants of abstract values
Summary:
Release Notes: None

One more invariant fix.

We used to use `object.property` instead of the `AbstractValue`'s serialized value for invariants. This causes both derived and normal references to be created for the AbstractValue which can cause the AbstractValue to be accessed before it is defined in some cases.
Closes https://github.com/facebook/prepack/pull/1467

Differential Revision: D7033556

Pulled By: cblappert

fbshipit-source-id: 68feeffb047e64fd4a39d9956e0166077405ee47