Error if ToPrimitive yields a topVal
If toString or valueOf yields a value that may not be either primitive or
object, then we don't know what to do, so we should fatal.

We used to incorrectly return that value.

An alternative could try to use a join here to try to figure out
at runtime which one was used.