Emit .call(this, ...) when emitting calls
We do something similar for CallExpression but not for this internal
operation.

This solves the issue with the ToPrimitive operation tracking abstract
calls to valueOf when valueOf is abstract.

However, this is unnecessary to emit this call since the binary operation
will force the internal operation anyway. Not sure how we should
model that.

We also have to leak the arguments when this internal call happens, just
like in the CallExpression evaluator.